Moved the enums into separate sub-folders
